+++ Sources/SentryCrash/Recording/Monitors/SentryCrashMonitorContext.h
@@ -201,7 +201,8 @@ typedef struct SentryCrash_MonitorContext {
         int parentProcessID;
         const char *deviceAppHash;
         const char *buildType;
-        uint64_t storageSize;
+        uint64_t totalStorageSize;
+        uint64_t freeStorageSize;
         uint64_t memorySize;
         uint64_t freeMemory;
         uint64_t usableMemory;