+++ Sources/SentryCrash/Recording/SentryCrash.m
@@ -258,7 +258,8 @@ getBasePath()
     COPY_PRIMITIVE(parentProcessID);
     COPY_STRING(deviceAppHash);
     COPY_STRING(buildType);
-    COPY_PRIMITIVE(storageSize);
+    COPY_PRIMITIVE(totalStorageSize);
+    COPY_PRIMITIVE(freeStorageSize);
     COPY_PRIMITIVE(memorySize);
     COPY_PRIMITIVE(freeMemory);
     COPY_PRIMITIVE(usableMemory);