+++ Tests/SentryTests/SentryCrash/TestSentryUIDeviceWrapper.swift
@@ -1,0 +1,30 @@
+import Sentry
+
+class TestSentryUIDeviceWrapper: SentryUIDeviceWrapper {
+#if os(iOS)
+    var internalOrientation = UIDeviceOrientation.portrait
+    var internalIsBatteryMonitoringEnabled = true
+    var internalBatteryLevel: Float = 0.6
+    var interalBatteryState = UIDevice.BatteryState.charging
+
+    override init() {
+        // Do nothing
+    }
+
+    override func orientation() -> UIDeviceOrientation {
+        return internalOrientation
+    }
+
+    override func isBatteryMonitoringEnabled() -> Bool {
+        return internalIsBatteryMonitoringEnabled
+    }
+
+    override func batteryLevel() -> Float {
+        return internalBatteryLevel
+    }
+
+    override func batteryState() -> UIDevice.BatteryState {
+        return interalBatteryState
+    }
+#endif
+}